Admin panel(image upload)

+++ modules/photogallery/modules/admin/controllers/CategoryController.php
@@ -35,6 +35,10 @@ class CategoryController extends Controller
      */
     public function actionIndex()
     {
+        if (Yii::$app->user->isGuest || Yii::$app->user->identity->username == "demo") {
+            return $this->goHome();
+        }
+
         $searchModel = new CategorySearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -52,6 +56,10 @@ class CategoryController extends Controller
      */
     public function actionView($id)
     {
+        if (Yii::$app->user->isGuest || Yii::$app->user->identity->username == "demo") {
+            return $this->goHome();
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -64,6 +72,10 @@ class CategoryController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->isGuest || Yii::$app->user->identity->username == "demo") {
+            return $this->goHome();
+        }
+
         $model = new Category();
 
         if ($model->load(Yii::$app->request->post())) {
@@ -89,6 +101,10 @@ class CategoryController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->isGuest || Yii::$app->user->identity->username == "demo") {
+            return $this->goHome();
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -109,9 +125,12 @@ class CategoryController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        if (Yii::$app->user->isGuest || Yii::$app->user->identity->username == "demo") {
+            return $this->goHome();
+        } else{
+            $this->findModel($id)->delete();
+            return $this->redirect(['index']);
+        }
     }
 
     /**
@@ -123,6 +142,10 @@ class CategoryController extends Controller
      */
     protected function findModel($id)
     {
+        if (Yii::$app->user->isGuest || Yii::$app->user->identity->username == "demo") {
+            return $this->goHome();
+        }
+        
         if (($model = Category::findOne($id)) !== null) {
             return $model;
         }
